Run the server in a Docker container
By design, doco is supposed to run multiple instances of the web
application in parallel. This can create issues with the port that the
application is using, since the different processes might try to bind to
the same port. Assigning and passing around a random port has proven to
be difficult, since we cannot as easily treat the application as a
blackbox.

Running the server as a Docker container provides a simple abstraction
to run multiple server instances on random ports. We can use Docker's
API to start a new container for each test case, query the port that has
been assigned to the container, and run the test against it.