Provide custom `Client` implementation to tests (#11)
The `Client` struct from fantoccini[^1] has been "extended" by creating
a custom struct that implements `Deref`. This client sets the base URL
to the internal IP address and port of the Docker container that runs
the server. This cleans up the interface for tests and removes two
arguments for test functions.

[^1]: https://crates.io/crates/fantoccini