Split logic to run tests into a separate test runner (#12)
The logic to run tests has been moved from the `Doco` struct into a
separate test runner. This unlocks the next step in the development of
doco, which is to create procedural macros that collect and execute the
tests. By splitting the functionality of `Doco`, we can use it as an API
that user's of the crate have to "implement" and use the logic in the
macros.