Initialize test runner before executing tests
The test runner previously performed the initialization of the test
environment as part of the `run` function. This causes a problem when
trying to use the runner in the future derive macro, which wraps the
test function and runs it in an asynchronous runtime. Since the
initialization is asynchronous as well, we would need to wrap that in
the runtime as well. With the current design, that would lead to nested
`block_on` calls, which would probably create some weird race
conditions.

Since the long-term plan is to have more hooks on the test runner (e.g.
`before` and `before_each`), it seems reasonable to already split the
initialization from the test and avoid the above problem.