Remove TestCase trait
The `TestCase` trait has been removed by tweaking the signature of the
test runner's `run` method. It now takes a synchronous method, which is
required by the future `doco::test` macro due to limitations in Rust's
type system. Concretely, it is not yet possible to pass pointers to
asynchronous functions around.

To still be able to use asynchronous code in a test, we are spawning a
new async runtime in a background thread for each test. While this is
less efficient than running all tests in the same thread, it still
allows us to pass a synchronous function pointer while allowing tests to
use async code. The macro will hide this complexity from users so that
they only need to annotate an asynchronous method and things will just
work magically by themselves.