Create derive macro that collects and runs tests (#21)
Two derive macros have been created that collect and run the tests.
Together, they can be used to set up a custom test harness and easily
implement end-to-end tests. The first macro, main, must be added to a
function that sets up and returns an instance of the Doco framework. The
second macro, test, can be used to annotate tests that will be run by
doco's test runner.

Sadly, it is not possible in Rust (yet) to pass async function pointers
around. So the macro needs to wrap the asynchronous test functions in a
synchronous function that executes the test in an isolated async runtime
in a background thread.

The macros do not provide a good developer experience yet. Namely, they
don't handle test failures gracefully and they don't provide the user
with any feedback on which test is currently being executed.