Print which test is currently running (#22)
As a first step towards a better developer experience, Doco now collects
and prints the name of the currently running test. Future improvements
should include printing the result of the test (`ok` or `failure`),
similar to how cargo's test runner does it, and including the module
path in the test name so that different modules can have the same name.