Support additional services (#35)
Most applications will probably require additional services to function,
like a database or a remote cache. These can now be defined as services
that are started before the server container is started.

The GitHub Actions workflows has been slightly adjusted to run the full
CI, including linters and formatters, on the examples.