Fix publishing the crates (#62)
A few different things failed when trying to publish the crates for the first
time using the GitHub Actions workflow. For one, crates.io requires crates to
have a valid license specified in their Cargo.toml. And Cargo does not support
running cargo publish on a workspace without specifying a specific crate, and it
requires a version for the internal doco-derive dependency.

Additionally, we also linked the README into the different crates to improve the
presentation of the crates on crates.io.